fix: widen the hub's public stream parameter to intptr_t + bump cpu/cuda pins (#50)
A `cudaStream_t` is a pointer and therefore 64-bit, but every public
`stream` parameter in the hub's device-dispatch functions was declared
`int`. `ff::dt_euclidean` and friends inspect the DLTensor's device and
forward the same argument to either `FF_CPU::` or `FF_CUDA::`:

    if (is_cuda) return FF_CUDA::dt_euclidean(inp_out, voxel_spacing, stream);
    else         return FF_CPU::dt_euclidean(inp_out, voxel_spacing, stream);

so both backends must agree on the parameter's type. Widens 114 parameter
declarations across the 8 modules to `intptr_t`, matching the now-widened
cpu/cuda dispatch layers. This is a signature-only change (verified with the
same non-signature-line grep used in the cpu-lib/cuda-lib PRs: empty).

Bumps `cpu` -> d969e2d (fastfields-cpu-lib#61) and `cuda` -> 33aa9b3
(fastfields-cuda-lib#35).

Verified with `make all CXX=clang++`: the full hub (cpu library + hub object
files + link) builds successfully end to end against the bumped pins,
producing libfastfields.so linked against libfastfields-cpu.so. This is a
real, complete build of the CPU path; the CUDA path is not linked by this
target (CUDA needs `USE_CUDA=1` and nvcc, exercised at the cuda-lib layer
instead).

Refs fastfields/fastfields-lib#4

+++ include/fastfields/api/pushpull.h
@@ -1,6 +1,7 @@
 #ifndef FF_LIB_PUSHPULL
 #define FF_LIB_PUSHPULL
 #include "dlpack.h"
+#include <cstdint>
 #include "defines.h"
 
 #ifndef FF_LIB_BOUND_SPLINE_T
@@ -61,7 +62,7 @@ void pull(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -79,7 +80,7 @@ void push(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -92,7 +93,7 @@ void count(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -110,7 +111,7 @@ void grad(
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
           bool       abs         = false,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /***********************************************************************
@@ -161,7 +162,7 @@ void pull_backward(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -182,7 +183,7 @@ void push_backward(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -199,7 +200,7 @@ void count_backward(
           int8_t     spline      = spline_t::Quadratic,
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 /**
@@ -223,7 +224,7 @@ void grad_backward(
           int8_t     bound       = bound_t::DCT2,
           int8_t     extrapolate = 1,
           bool       abs         = false,
-          int        stream      = 0
+          intptr_t   stream      = 0
 );
 
 FF_NAMESPACE_END(FF)
