+++ include/fastfields/api/reg_flow.h
@@ -1,6 +1,7 @@
 #ifndef FF_LIB_REG_FLOW
 #define FF_LIB_REG_FLOW
 #include "dlpack.h"
+#include <cstdint>
 #include "defines.h"
 
 #ifndef FF_LIB_BOUND_SPLINE_T
@@ -72,7 +73,7 @@ void flow_matvec(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -90,7 +91,7 @@ void flow_addmatvec_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -108,7 +109,7 @@ void flow_submatvec_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -125,7 +126,7 @@ void flow_diag(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -149,7 +150,7 @@ void flow_kernel(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -174,7 +175,7 @@ void flow_adddiag_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -192,7 +193,7 @@ void flow_subdiag_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -210,7 +211,7 @@ void flow_addkernel_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -228,7 +229,7 @@ void flow_subkernel_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 void flow_relax(
@@ -244,7 +245,7 @@ void flow_relax(
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -270,7 +271,7 @@ void flow_forward(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -295,7 +296,7 @@ void flow_precond(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -316,7 +317,7 @@ void flow_precond_(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -348,7 +349,7 @@ void flow_matvec_rls(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -366,7 +367,7 @@ void flow_diag_rls(
           double     div       = 0.0,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -392,7 +393,7 @@ void flow_relax_rls(
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 FF_NAMESPACE_END(FF)
